Move hooks and libs folder. Create utils folder.
