+++ src/components/app/App.js
@@ -4,9 +4,9 @@ import Footer from '../footer/Footer'
 import Home from '../home/Home'
 import ActivitiesList from '../ActivitiesList/ActivitiesList'
 import CreateActivityForm from '../createActivityForm/CreateActivityForm'
-import loadFromLocal from '../lib/loadFromLocal'
-import saveToLocal from '../lib/saveToLocal'
-import filterActivities from '../lib/filterActivities'
+import loadFromLocal from '../../lib/loadFromLocal'
+import saveToLocal from '../../lib/saveToLocal'
+import filterActivities from '../../utils/filterActivities'
 import styled from 'styled-components/macro'
 import { ThemeProvider } from 'styled-components'
 import { theme } from '../../theme'
