+++ src/components/header/Header.js
@@ -5,7 +5,7 @@ import logo from '../../assets/images/logo.svg'
 import styled from 'styled-components/macro'
 import { useState, useRef } from 'react'
 import { NavLink } from 'react-router-dom'
-import useClickOutside from '../hooks/useClickOutside'
+import useClickOutside from '../../hooks/useClickOutside'
 
 export default function Header({ searchTerm, onChange }) {
   const [open, setOpen] = useState(false)
