+++ src/lib/loadFromLocal.js
